Implement and use `magit-can-use-p' and `magit-can-use-assert'
This eliminates explicit tests against the git version in favour of
testing for the git features being used, by name.

Supporting multiple feature arguments further reduces the need to
always manually track and specify the feature with the highest version
dependency, when multiple different features are in play.  Changes
which add or remove a dependency on any given git feature then need
only update the named list of features being depended on, without
worrying what the highest git version dependency will be afterwards.

The supported features are cached on a per-repository basis, so that
the available feature set for remote repositories is always in sync
with the remote git version.  This reduces the number of calls to
`magit-git-version'.